Backend/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pycodestyle==2.13.0
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
django-allauth==65.7.0
cryptography
requests
requests_oauthlib